+++ java/client/src/main/java/org/signal/libsignal/net/RateLimitChallengeException.kt
@@ -6,6 +6,7 @@
 package org.signal.libsignal.net
 
 import org.signal.libsignal.internal.CalledFromNative
+import java.time.Duration
 import java.util.EnumSet
 
 /**
@@ -17,10 +18,30 @@ import java.util.EnumSet
 public class RateLimitChallengeException : ChatServiceException {
   public val token: String
   public val options: Set<ChallengeOption>
+  public val retryLater: Duration?
 
-  @CalledFromNative
-  public constructor(message: String, token: String, options: Array<ChallengeOption>) : super(message) {
+  public constructor(
+    message: String,
+    token: String,
+    options: Array<ChallengeOption>,
+    retryLater: Duration?,
+  ) : super(message) {
     this.token = token
     this.options = EnumSet.copyOf(options.asList())
+    this.retryLater = retryLater
+  }
+
+  @CalledFromNative
+  internal constructor(
+    message: String,
+    token: String,
+    options: Array<ChallengeOption>,
+    retryLater: Long,
+  ) : this(
+    message,
+    token,
+    options,
+    if (retryLater < 0) null else Duration.ofSeconds(retryLater),
+  ) {
   }
 }