+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
 v0.90.0
 
+- Add support for a retry later duration in rate limit responses
+
 - keytrans: Simplify the top-level API
 
 - Use CDSI enclave 3a1ac5e5 in staging.